+++ src/main/java/org/folio/linked/data/mapper/dto/monograph/instance/InstanceMapperUnit.java
@@ -83,14 +83,14 @@ public class InstanceMapperUnit extends TopResourceMapperUnit {
   private final HashService hashService;
 
   @Override
-  public <P> P toDto(Resource source, P parentDto, Resource parentResource) {
+  public <P> P toDto(Resource resourceToConvert, P parentDto, ResourceMappingContext context) {
     if (parentDto instanceof ResourceResponseDto resourceDto) {
-      var instance = coreMapper.toDtoWithEdges(source, InstanceResponse.class, false);
-      instance.setId(String.valueOf(source.getId()));
-      ofNullable(source.getFolioMetadata())
+      var instance = coreMapper.toDtoWithEdges(resourceToConvert, InstanceResponse.class, false);
+      instance.setId(String.valueOf(resourceToConvert.getId()));
+      ofNullable(resourceToConvert.getFolioMetadata())
         .map(folioMetadataMapper::toDto)
         .ifPresent(instance::setFolioMetadata);
-      ofNullable(source.getDoc())
+      ofNullable(resourceToConvert.getDoc())
         .ifPresent(doc -> instance.setNotes(noteMapper.toNotes(doc, SUPPORTED_NOTES)));
       resourceDto.resource(new InstanceResponseField().instance(instance));
     }
@@ -117,7 +117,7 @@ public class InstanceMapperUnit extends TopResourceMapperUnit {
     coreMapper.addOutgoingEdges(instance, InstanceRequest.class, instanceDto.getCarrier(), CARRIER);
     coreMapper.addOutgoingEdges(instance, InstanceRequest.class, instanceDto.getCopyright(), COPYRIGHT);
     coreMapper.addOutgoingEdges(instance, InstanceRequest.class, instanceDto.getWorkReference(), INSTANTIATES);
-    coreMapper.addOutgoingEdges(instance, InstanceRequest.class, instanceDto.getExtentV2(), EXTENT);
+    coreMapper.addOutgoingEdges(instance, InstanceRequest.class, instanceDto.getExtent(), EXTENT);
     coreMapper.addOutgoingEdges(instance, InstanceRequest.class, standardBookFormats(instanceDto), BOOK_FORMAT);
     instance.setFolioMetadata(new FolioMetadata(instance).setSource(LINKED_DATA));
     instance.setId(hashService.hash(instance));
@@ -126,7 +126,6 @@ public class InstanceMapperUnit extends TopResourceMapperUnit {
 
   private JsonNode getDoc(InstanceRequest dto) {
     var map = new HashMap<String, List<String>>();
-    putProperty(map, PropertyDictionary.EXTENT, dto.getExtent());
     putProperty(map, DIMENSIONS, dto.getDimensions());
     putProperty(map, EDITION, dto.getEdition());
     putProperty(map, PROJECTED_PROVISION_DATE, dto.getProjectProvisionDate());